LayoutTests/platform/glib/imported/w3c/web-platform-tests/streams/piping/multiple-propagation.any-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: reader released lock
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: reader released lock
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!
CONSOLE MESSAGE: Unhandled Promise Rejection: error1: error1!

Harness Error (FAIL), message = Unhandled rejection: error1!

PASS Piping from an errored readable stream to an erroring writable stream
PASS Piping from an errored readable stream to an errored writable stream
PASS Piping from an errored readable stream to an erroring writable stream; preventAbort = true
PASS Piping from an errored readable stream to an errored writable stream; preventAbort = true
PASS Piping from an errored readable stream to a closing writable stream
PASS Piping from an errored readable stream to a closed writable stream
PASS Piping from a closed readable stream to an erroring writable stream
PASS Piping from a closed readable stream to an errored writable stream
PASS Piping from a closed readable stream to a closed writable stream
